Adding record proxy type
